deta/requirements.txt
flask
deta
python-dotenv
flask_cors